transitioning files to be more modular and take advantage of ES6 features more often.
