+++ 2.0playground/index.html
@@ -154,7 +154,7 @@
         
     </section>
 
-    <section id="qPrompt">
+    <section id="qPrompt" class="hide">
         <div class="modal-cont">
             <button type="button" id="closeModal">x</button>
             <span id="modalSong"></span>
@@ -171,4 +171,4 @@
 
 </html>
 
-<script type="module" src="index.js"></script>
+<!-- <script type="module" src="index.js"></script> -->
