+++ 2.0playground/styles.css
@@ -1,7 +1,190 @@
 .body-mod {
     display: flex;
-    /* flex-direction: column; */
+    flex-direction: column;
     justify-content: center;
     margin: 10px auto;
     background-color: azure;
+}
+
+#wellHeyThere {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    padding: 15px 0;
+    font-size: 3vmin;
+    font-family: serif;
+    background-color: floralwhite;
+    text-shadow: 0 2px 3px darkgrey;
+}
+
+header {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+/*  */
+/* Audio player */
+/*  */
+#player {
+    width: 100%;
+    margin: 0;
+    padding: 0;
+    border-bottom: 5px solid white;
+}
+audio::-webkit-media-controls-panel {
+    background-color: rgb(203 249 236);
+}
+audio::-webkit-media-controls-enclosure {
+    border-radius: 0px;
+}
+audio::-webkit-media-controls-timeline-container {
+    color:green;
+    border: 1px solid black;
+}
+audio::-webkit-media-controls-current-time-display, audio::-webkit-media-controls-time-remaining-display {
+    color:black;
+    font-family: serif;
+    padding: 0 3px;
+    margin: auto;
+}
+audio::-webkit-media-controls-timeline {
+    color: black;
+    background-color: white;
+    border-radius: 50px;
+    padding: 20px;
+    margin-top:10px;
+}
+
+/*  */
+/* Song info and Queue containers */
+/*  */
+#trackController {
+    display: flex;
+    height:200px;
+    background-color:rgb(203 249 236);
+    overflow: hidden;
+}
+
+#anotherPlayer {
+    display: grid;
+    width: 50%;
+    grid-template-columns: 1fr 2fr 3fr;
+    padding-left: 5px;
+}
+
+#qTools {
+    display: flex;
+    flex-direction: column;
+    width: 50%;
+    padding-left: 5px;
+    overflow-y: scroll;
+}
+
+
+.option-buttons {
+    width: 100%;
+}
+
+
+/*  */
+/* Now playing container */
+/*  */
+#albumPic {
+    height: 100px;
+    width: 100px;
+    margin: 10px;
+}
+
+#nowPlayingInfo {
+    display: flex;
+    flex-direction: column;
+    padding: 5px 0px;
+}
+
+#artistName, #trackName, #albumName {
+    margin: 5px;
+    border-bottom: 1px solid #9ebb00;
+}
+
+#artistName {
+    font-size: 20px;
+}
+#trackName {
+    font-style: oblique;
+    font-weight: bold;
+}
+#albumName {
+    color:rgb(78, 78, 78);
+}
+
+
+/*  */
+/* Queue List Container */
+/*  */
+.queue-tip {
+    color: grey;
+}
+
+#queueList {
+    flex-direction: column;
+    width: 100%;
+}
+
+#queueList > button {
+    background-color: orange;
+    border: none;
+    border-left: 15px dotted #000000;
+    cursor: pointer;
+}
+#queueList > button:first-child::before {
+    content: " up next! ~~ ";
+    font-weight: bold;
+}
+
+#queueList > button:hover {
+    border-bottom: 2px solid #c1c2c3;
+    border-top: 1px solid #c1c2c3;
+}
+
+
+/*  */
+/* Play Pause Next Queue Buttons */
+/*  */
+.controlBtns {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    margin: 0 40px;
+}
+
+.controlBtns > button {
+    padding: 10px;
+    height: 60px;
+    width: 60px;
+    margin: auto;
+}
+
+#nextTrack {
+    background-color: darkorange;
+}
+
+.glyphicon {
+    font-size: 25px;
+}
+
+
+/*  */
+.song-row {
+    display: grid;
+    grid-template-columns: .3fr 2fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr;
+}
+
+.song-row > span {
+    margin-right: 20px;
+}
+
+
+
+.hide {
+    display: none;
 }
