+++ assets/js/sketches.js
@@ -1,0 +1,929 @@
+
+export const sketches = [
+    {
+        title: 'BCC',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'November 21st, 2021',
+        artLink: './assets/bcc.png',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/bcc.png',
+        trackList: [
+            {
+                track: 'Hotmess',
+                length: '3:04',
+                url: 'bcc-hotmess',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Atmosphere (interlude)',
+                length: '0:35',
+                url: 'bcc-atmosphere',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Amanda',
+                length: '1:47',
+                url: 'bcc-amanda',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Dare2Bare',
+                length: '2:35',
+                url: 'bcc-dare2bare',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Quarantine (but its ok)',
+                length: '3:01',
+                url: 'bcc-quarantine',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Stickyhold',
+                length: '1:27',
+                url: 'bcc-stickyhold',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Stargate (reprise)',
+                length: '1:26',
+                url: 'bcc-stargate',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Rap',
+                length: '0:54',
+                url: 'bcc-rap',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: '27th&girard',
+                length: '4:18',
+                url: 'bcc-girard',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Almostinato',
+                length: '2:34',
+                url: 'bcc-it-is-forever',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: '5d',
+                length: '2:08',
+                url: 'bcc-5d',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Probably Wrong',
+                length: '2:58',
+                url: 'bcc-probably',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'S. Balachander (bftb mx)',
+                length: '4:59',
+                url: 'bcc-sbala',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Boc (interlude)',
+                length: '1:22',
+                url: 'bcc-boc',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Toothpicks',
+                length: '2:31',
+                url: 'bcc-toothpicks',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Pilotdrone',
+                length: '2:27',
+                url: 'bcc-pilotdrone',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Faultline',
+                length: '4:54',
+                url: 'bcc-faultline',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+        ]
+    },
+    {
+        title: 'S&OFV1',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'August 28th, 2017',
+        artLink: './assets/s&ofv1.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv1.jpg',
+        trackList: [
+            {
+                track: 'Quarantine (but its ok)',
+                length: '2:56',
+                url: 'quarantine',
+                fileName: 'pilot',
+                daw: 'LogicX',
+                date: "06-25-17",
+                started: '11:35 PM',
+            },
+            {
+                track: '5d',
+                length: '3:01',
+                url: '5d',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Sticky Hold',
+                length: '1:27',
+                url: 'sticky-hold',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Probably Wrong',
+                length: '3:16',
+                url: 'probably-wrong',
+                fileName: 'smoov',
+                daw: 'LogicX',
+                date: '06-08-17',
+                started: '2:46 AM',
+            },
+            {
+                track: 'Stargate',
+                length: '1:13',
+                url: 'stargate',
+                fileName: 'stargate',
+                daw: 'LogicX',
+                date: '--',
+                started: '1:21 AM',
+            },
+            {
+                track: 'Rap (1896 mix)',
+                length: '0:56',
+                url: 'rap-1896',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Dare 2 Bare',
+                length: '2:53',
+                url: 'dare-2-bare',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+
+        ]
+    },
+    {
+        title: 'S&OFV2',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/s&ofv2.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv2.jpg',
+        trackList: [
+            {
+                track: 'Amanda Alarm (wake up hunny)',
+                length: '1:56',
+                url: 'amanda-alarm',
+                fileName: 'bell-bottom-chords',
+                daw: 'LogicX',
+                date: '06-17-17',
+                started: '8:13 PM',
+            },
+            {
+                track: 'Doogood',
+                length: '1:32',
+                url: 'dogood',
+                fileName: 'dogood',
+                daw: 'LogicX',
+                date: '08-07-18',
+                started: '10:31 PM',
+            },
+            {
+                track: 'Exploitable Resources',
+                length: '1:24',
+                url: 'resources',
+                fileName: 'dougle',
+                daw: 'LogicX',
+                date: '10-06-18',
+                started: '11:00 PM',
+            },
+            {
+                track: 'Looking at a Screen',
+                length: '2:57',
+                url: 'looking-at-a-screen',
+                fileName: 'poppinggm',
+                daw: 'LogicX',
+                date: '08-12-17',
+                started: '11:16 PM',
+            },
+            {
+                track: 'Woodshed',
+                length: "0:52",
+                url: 'woodshed',
+                fileName: 'broot',
+                daw: 'FLstudio',
+                date: '05-07-18',
+                started: '2:44 PM',
+            },
+            {
+                track: 'Gimp',
+                length: '2:31',
+                url: 'gimp',
+                fileName: 'future stranger',
+                daw: 'LogicX',
+                date: '10-09-18',
+                started: '11:44 PM',
+            },
+            {
+                track: 'Amanda Morph (gettin late)',
+                length: '1:46',
+                url: 'amanda-morph',
+                fileName: 'bell-bottom-chords',
+                daw: 'LogicX',
+                date: '06-17-17',
+                started: '8:13 PM',
+            },
+        ]
+    },
+    {
+        title: 'S&OFV3',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/s&ofv3.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv3.jpg',
+        trackList: [
+            {
+                track: 'Sara / Tony Tonin',
+                length: '2:22',
+                url: 'sara-tonin',
+                fileName: 'rottunspahetti',
+                daw: 'LogicX',
+                date: '08-11-17',
+                started: '12:03 AM',
+            },
+            {
+                track: 'Sweet Dreams',
+                length: '2:35',
+                url: 'sweet-dreams',
+                fileName: 'sleep dreams',
+                daw: 'LogicX',
+                date: '05-15-18',
+                started: '8:45 PM',
+            },
+            {
+                track: 'Bailey melody 07',
+                length: "0:59",
+                url: 'bailey-melody',
+                fileName: 'bounce12telly',
+                daw: 'LogicX',
+                date: '09-25-17',
+                started: '12:08 AM',
+            },
+            {
+                track: '(I am aware of the) black box',
+                length: "1:50",
+                url: 'black-box',
+                fileName: 'satch',
+                daw: 'LogicX',
+                date: '10-05-17',
+                started: '2:50 AM',
+            },
+            {
+                track: 'Oh techre!',
+                length: "2:09",
+                url: 'oh-techre',
+                fileName: 'techre',
+                daw: 'LogicX',
+                date: '06-21-17',
+                started: '12:42 AM',
+            },
+            {
+                track: '1800harp',
+                length: "2:55",
+                url: '1800harp',
+                fileName: '1800harp',
+                daw: 'LogicX',
+                date: '04-16-17',
+                started: '3:17 PM',
+            },
+            {
+                track: 'Go to feather',
+                length: "1:28",
+                url: 'go2feather',
+                fileName: 'skum****',
+                daw: 'LogicX',
+                date: '07-27-17',
+                started: '1:42 AM',
+            },
+        ]
+    },
+    {
+        title: 'S&OFV4',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/s&ofv4.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv4.jpg',
+        trackList: [
+            {
+                track: 'Hello grommet',
+                length: "3:10",
+                url: 'hello-grommet',
+                fileName: 'project99',
+                daw: 'LogicX',
+                date: '06-27-17',
+                started: '4:38 PM',
+            },
+            {
+                track: 'Froggy',
+                length: "1:10",
+                url: 'froggy',
+                fileName: 'cycle psycho',
+                daw: 'LogicX',
+                date: '06-02-18',
+                started: '4:15 PM',
+            },
+            {
+                track: 'Live from the polls',
+                length: "1:08",
+                url: 'polls',
+                fileName: 'flowyhits',
+                daw: 'LogicX',
+                date: '06-19-17',
+                started: '4:44 PM',
+            },
+            {
+                track: 'BBBB',
+                length: "1:43",
+                url: 'bbbb',
+                fileName: 'wipsmalbas',
+                daw: 'LogicX',
+                date: '06-25-17',
+                started: '5:56 PM',
+            },
+            {
+                track: 'Case of the cookies',
+                length: "1:45",
+                url: 'cookies',
+                fileName: 'apple tree',
+                daw: 'LogicX',
+                date: '01-10-18',
+                started: '4:57 PM',
+            },
+            {
+                track: 'Silly Singing silicon',
+                length: "2:35",
+                url: 'silly-singing',
+                fileName: 'flunk',
+                daw: 'LogicX',
+                date: '08-28-18',
+                started: '6:22 PM',
+            },
+            {
+                track: 'Is it forever?',
+                length: "2:24",
+                url: 'is-it-forever',
+                fileName: 'flips',
+                daw: 'LogicX',
+                date: '09-13-17',
+                started: '5:02 PM',
+            },
+        ]
+    },
+    {
+        title: 'S&OFV5',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/s&ofv5.png',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv5.png',
+        trackList: [
+            {
+                track: 'Are Those Thorns?',
+                length: "1:02",
+                url: 'are-those-thorns',
+                fileName: 'dr insta',
+                daw: 'FLstudio',
+                date: '05-04-19',
+                started: '4:36 PM',
+            },
+            {
+                track: 'Barrixtown',
+                length: "1:10",
+                url: 'barrixtown',
+                fileName: 'BONKUS',
+                daw: 'FLstudio',
+                date: '08-05-19',
+                started: '12:43 PM',
+            },
+            {
+                track: 'Childs Loop',
+                length: "1:48",
+                url: 'childs-loop',
+                fileName: 'dingus',
+                daw: 'FLstudio',
+                date: '10-11-18',
+                started: '5:29 PM',
+            },
+            {
+                track: 'Cleverly Loop',
+                length: "1:43",
+                url: 'clever',
+                fileName: 'clever',
+                daw: 'FLstudio',
+                date: '11-10-18',
+                started: '5:29 PM',
+            },
+            {
+                track: 'Chix Loop',
+                length: "2:00",
+                url: 'chix',
+                fileName: 'chix',
+                daw: 'FLstudio',
+                date: '05-17-19',
+                started: '11:12 PM',
+            },
+            {
+                track: 'FL Disaster Zone',
+                length: "1:53",
+                url: 'fl-disaster',
+                fileName: 'cut90butwe',
+                daw: 'FLstudio',
+                date: '10-02-18',
+                started: '1:25 PM',
+            },
+            {
+                track: 'Crystal Envelope / Return2sender',
+                length: "4:14",
+                url: 'crystal-env',
+                fileName: 'dueling pianos',
+                daw: 'FLstudio',
+                date: '05-05-19',
+                started: '5:50 PM',
+            },
+        ]
+    },
+    {
+        title: 'S&OFV6',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/oops.png',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/oops.png',
+        trackList: [
+            {
+                track: 'Parks',
+                length: "0:20",
+                url: 'parks',
+                fileName: 'parks',
+                daw: 'FLstudio',
+                date: '04-18-19',
+                started: '1:43 PM',
+            },
+            {
+                track: 'Playing With My Hair',
+                length: "1:27",
+                url: 'playing-w-my-hair',
+                fileName: 'clair',
+                daw: 'FLstudio',
+                date: '04-13-19',
+                started: '9:32 PM',
+            },
+            {
+                track: 'Rust In Peace',
+                length: "1:43",
+                url: 'rust-in-peace',
+                fileName: 'disgus rstie',
+                daw: 'FLstudio',
+                date: '09-28-18',
+                started: '8:12 PM',
+            },
+            {
+                track: 'Kicking Space',
+                length: "1:18",
+                url: 'kicking-space',
+                fileName: 'creating space',
+                daw: 'FLstudio',
+                date: '04-13-19',
+                started: '10:57 AM',
+            },
+            {
+                track: 'Let It Fly',
+                length: "1:47",
+                url: 'let-it-fly',
+                fileName: 'destroying space',
+                daw: 'FLstudio',
+                date: '04-20-19',
+                started: '10:04 AM',
+            },
+            {
+                track: 'Liptstick Halo',
+                length: "1:31",
+                url: 'liptstick-halo',
+                fileName: 'bop',
+                daw: 'FLstudio',
+                date: '07-28-18',
+                started: '10:32 AM',
+            },
+            {
+                track: 'It Is Forever',
+                length: '?:??',
+                url: 'it-is-forever',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+
+            },
+        ]
+    },
+    {
+        title: 'S&OFV7',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/s&ofv7.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv7.jpg',
+        trackList: [
+            {
+                track: '404',
+                length: '1:00',
+                url: '404',
+                fileName: 'hat slide shift',
+                daw: 'FLstudio',
+                date: '08-24-18',
+                started: '9:20 PM',
+            },
+            {
+                track: 'Purple!',
+                length: '0:45',
+                url: 'purple',
+                fileName: 'hot spooky mess',
+                daw: 'FLstudio',
+                date: '05-31-19',
+                started: '11:54 PM',
+            },
+            {
+                track: 'Bubble Boy',
+                length: '1:54',
+                url: 'bubble-boy',
+                fileName: 'saw distort',
+                daw: 'FLstudio',
+                date: '07-05-18',
+                started: '1:48 PM',
+            },
+            {
+                track: 'Flawz',
+                length: '2:40',
+                url: 'flawz',
+                fileName: 'flaws',
+                daw: 'FLstudio',
+                date: '04-02-19',
+                started: '3:30PM',
+            },
+            {
+                track: 'Gabby',
+                length: '0:40',
+                url: 'gabby',
+                fileName: 'gabby',
+                daw: 'FLstudio',
+                date: '03-13-19',
+                started: '5:06 PM',
+            },
+            {
+                track: 'Midi Compactor',
+                length: '1:19',
+                url: 'midi-compactor',
+                fileName: 'empty',
+                daw: 'FLstudio',
+                date: '07-17-19',
+                started: '5:27 PM',
+            },
+            {
+                track: 'Rip Loop',
+                length: '3:43',
+                url: 'rip-arp',
+                fileName: 'protofunk',
+                daw: 'FLstudio',
+                date: '07-02-19',
+                started: '1:00 PM',
+            },
+        ]
+    },
+    {
+        title: 'S&OFV8',
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/s&ofv8.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv8.jpg',
+        trackList: [
+            {
+                track: 'Rogue on 1-1',
+                length: '1:46',
+                url: 'rogue-on-1-1',
+                fileName: 'jumbles',
+                daw: 'FLstudio',
+                date: '05-03-19',
+                started: '5:12 PM',
+            },
+            {
+                track: 'Sea of Drones',
+                length: '3:34',
+                url: 'sea-of-drones',
+                fileName: 'sea of drones',
+                daw: 'FLstudio',
+                date: '06-25-19',
+                started: '10:42 PM',
+            },
+            {
+                track: 'Sine Language',
+                length: '1:42',
+                url: 'sine-lang',
+                fileName: 'drane',
+                daw: 'FLstudio',
+                date: '10-26-18',
+                started: '10:22 PM',
+            },
+            {
+                track: 'SSX Loop',
+                length: '1:08',
+                url: 'ssx-loop',
+                fileName: 'return',
+                daw: 'FLstudio',
+                date: '09-22-18',
+                started: '5:24 PM',
+            },
+            {
+                track: 'The Drs In.',
+                length: '1:18',
+                url: 'the-doctors-in',
+                fileName: 'immortal sounds',
+                daw: 'FLstudio',
+                date: '05-10-19',
+                started: '3:28 PM',
+            },
+            {
+                track: 'girard (noisey)',
+                length: '4:47',
+                url: 'girard-noisy',
+                fileName: '--',
+                daw: '--',
+                date: '--',
+                started: '--',
+            },
+        ]
+    },
+    
+    {
+        title: "IB",
+        artist: 'con',
+        longArtist: 'con',
+        releaseDate: 'Avril 4th, 2014',
+        artLink: './assets/ib.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/ib.jpg',
+        trackList: [
+            {
+                track: 'Ferox Recieved',
+                length: "3:17",
+                url: 'ferox',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Tsalal',
+                length: "1:50",
+                url: 'tsalal',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Anxious...',
+                length: "4:35",
+                url: 'anxious-oblast',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Knavish Drone (cut)',
+                length: "16:07",
+                url: 'knavish-drone',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+        ]
+    },
+    {
+        title: "Redactor",
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/redact.png',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/redact.png',
+        trackList: [
+            {
+                track: 'In a room',
+                length: "12:45",
+                url: 'on-line',
+                fileName: '--',
+                daw: 'GB/TAS',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Untitled (miccheck)',
+                length: '14:19',
+                url: 'on-line1',
+                fileName: '--',
+                daw: 'GB/TAS',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Untitled (miccheck1)',
+                length: '6:13',
+                url: 'on-line2',
+                fileName: '--',
+                daw: 'GB/TAS',
+                date: '--',
+                started: '--',
+            },
+        ]
+    },
+    {
+        title: "ECT",
+        artist: 'L. Mauro',
+        longArtist: 'Lukasz Mauro',
+        releaseDate: 'N/A',
+        artLink: './assets/sketches.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/ect.png',
+        trackList: [
+            {
+                track: 'This is forever (demo)',
+                length: "2:36",
+                url: 'this-is-forever-v1',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'HM/FL (demo)',
+                length: "7:57",
+                url: 'hmfl-demo',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'HM/FL (survivor)',
+                length: '7:57',
+                url: 'hmfl(survivor)',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+        ]
+    },
+    {
+        title: "studs",
+        artist: 'S. Finder',
+        longArtist: 'Stud Finder',
+        releaseDate: 'N/A',
+        artLink: './assets/studs.jpg',
+        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/studs.jpg',
+        trackList: [
+            {
+                track: 'Boiler Room',
+                length: "1:43",
+                url: 'boiler-room',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+            {
+                track: 'Chiminy Harker',
+                length: "0:43",
+                url: 'chiminy',
+                fileName: '--',
+                daw: 'FLstudio',
+                date: '--',
+                started: '--',
+            },
+        ]
+    },
+    {
+        title: "BRSSF Pods",
+        artist: "Burner B",
+        longArtist: "DJ Burna B",
+        releaseDate: "September 30th, 2018",
+        artLink: "./assets/pods.jpg",
+        webLink: "https://burnthis.space/assets/images/pods.jpg",
+        trackList: [
+            {
+                track: 'Soundscape Source Feed vol. 1',
+                length: "33:33",
+                url: 'pod1-198',
+                fileName: 'brssfv1',
+                daw: 'FLstudio',
+                date: '09-30-18',
+                started: '--',
+            },
+            {
+                track: 'Soundscape Source Feed vol. 2',
+                length: "55:08",
+                url: 'pod2-112',
+                fileName: 'brssfv2',
+                daw: 'FLstudio',
+                date: '12-26-18',
+                started: '--',
+            },
+            {
+                track: 'Soundscape Source Feed vol. 3',
+                length: "40:27",
+                url: 'pod3-160',
+                fileName: 'brssfv3',
+                daw: 'FLstudio',
+                date: '03-01-19',
+                started: '--',
+            }
+        ]
+    }
+]
