--- assets/scripts/scripts.js
@@ -1,1853 +1,0 @@
-
-const sketches = [
-    {
-        title: 'BCC',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'November 21st, 2021',
-        artLink: './assets/bcc.png',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/bcc.png',
-        trackList: [
-            {
-                track: 'Hotmess',
-                length: '3:04',
-                url: 'bcc-hotmess',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Atmosphere (interlude)',
-                length: '0:35',
-                url: 'bcc-atmosphere',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Amanda',
-                length: '1:47',
-                url: 'bcc-amanda',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Dare2Bare',
-                length: '2:35',
-                url: 'bcc-dare2bare',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Quarantine (but its ok)',
-                length: '3:01',
-                url: 'bcc-quarantine',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Stickyhold',
-                length: '1:27',
-                url: 'bcc-stickyhold',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Stargate (reprise)',
-                length: '1:26',
-                url: 'bcc-stargate',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Rap',
-                length: '0:54',
-                url: 'bcc-rap',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: '27th&girard',
-                length: '4:18',
-                url: 'bcc-girard',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Almostinato',
-                length: '2:34',
-                url: 'bcc-it-is-forever',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: '5d',
-                length: '2:08',
-                url: 'bcc-5d',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Probably Wrong',
-                length: '2:58',
-                url: 'bcc-probably',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'S. Balachander (bftb mx)',
-                length: '4:59',
-                url: 'bcc-sbala',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Boc (interlude)',
-                length: '1:22',
-                url: 'bcc-boc',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Toothpicks',
-                length: '2:31',
-                url: 'bcc-toothpicks',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Pilotdrone',
-                length: '2:27',
-                url: 'bcc-pilotdrone',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Faultline',
-                length: '4:54',
-                url: 'bcc-faultline',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-        ]
-    },
-    {
-        title: 'S&OFV1',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'August 28th, 2017',
-        artLink: './assets/s&ofv1.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv1.jpg',
-        trackList: [
-            {
-                track: 'Quarantine (but its ok)',
-                length: '2:56',
-                url: 'quarantine',
-                fileName: 'pilot',
-                daw: 'LogicX',
-                date: "06-25-17",
-                started: '11:35 PM',
-            },
-            {
-                track: '5d',
-                length: '3:01',
-                url: '5d',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Sticky Hold',
-                length: '1:27',
-                url: 'sticky-hold',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Probably Wrong',
-                length: '3:16',
-                url: 'probably-wrong',
-                fileName: 'smoov',
-                daw: 'LogicX',
-                date: '06-08-17',
-                started: '2:46 AM',
-            },
-            {
-                track: 'Stargate',
-                length: '1:13',
-                url: 'stargate',
-                fileName: 'stargate',
-                daw: 'LogicX',
-                date: '--',
-                started: '1:21 AM',
-            },
-            {
-                track: 'Rap (1896 mix)',
-                length: '0:56',
-                url: 'rap-1896',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Dare 2 Bare',
-                length: '2:53',
-                url: 'dare-2-bare',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-
-        ]
-    },
-    {
-        title: 'S&OFV2',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/s&ofv2.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv2.jpg',
-        trackList: [
-            {
-                track: 'Amanda Alarm (wake up hunny)',
-                length: '1:56',
-                url: 'amanda-alarm',
-                fileName: 'bell-bottom-chords',
-                daw: 'LogicX',
-                date: '06-17-17',
-                started: '8:13 PM',
-            },
-            {
-                track: 'Doogood',
-                length: '1:32',
-                url: 'dogood',
-                fileName: 'dogood',
-                daw: 'LogicX',
-                date: '08-07-18',
-                started: '10:31 PM',
-            },
-            {
-                track: 'Exploitable Resources',
-                length: '1:24',
-                url: 'resources',
-                fileName: 'dougle',
-                daw: 'LogicX',
-                date: '10-06-18',
-                started: '11:00 PM',
-            },
-            {
-                track: 'Looking at a Screen',
-                length: '2:57',
-                url: 'looking-at-a-screen',
-                fileName: 'poppinggm',
-                daw: 'LogicX',
-                date: '08-12-17',
-                started: '11:16 PM',
-            },
-            {
-                track: 'Woodshed',
-                length: "0:52",
-                url: 'woodshed',
-                fileName: 'broot',
-                daw: 'FLstudio',
-                date: '05-07-18',
-                started: '2:44 PM',
-            },
-            {
-                track: 'Gimp',
-                length: '2:31',
-                url: 'gimp',
-                fileName: 'future stranger',
-                daw: 'LogicX',
-                date: '10-09-18',
-                started: '11:44 PM',
-            },
-            {
-                track: 'Amanda Morph (gettin late)',
-                length: '1:46',
-                url: 'amanda-morph',
-                fileName: 'bell-bottom-chords',
-                daw: 'LogicX',
-                date: '06-17-17',
-                started: '8:13 PM',
-            },
-        ]
-    },
-    {
-        title: 'S&OFV3',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/s&ofv3.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv3.jpg',
-        trackList: [
-            {
-                track: 'Sara / Tony Tonin',
-                length: '2:22',
-                url: 'sara-tonin',
-                fileName: 'rottunspahetti',
-                daw: 'LogicX',
-                date: '08-11-17',
-                started: '12:03 AM',
-            },
-            {
-                track: 'Sweet Dreams',
-                length: '2:35',
-                url: 'sweet-dreams',
-                fileName: 'sleep dreams',
-                daw: 'LogicX',
-                date: '05-15-18',
-                started: '8:45 PM',
-            },
-            {
-                track: 'Bailey melody 07',
-                length: "0:59",
-                url: 'bailey-melody',
-                fileName: 'bounce12telly',
-                daw: 'LogicX',
-                date: '09-25-17',
-                started: '12:08 AM',
-            },
-            {
-                track: '(I am aware of the) black box',
-                length: "1:50",
-                url: 'black-box',
-                fileName: 'satch',
-                daw: 'LogicX',
-                date: '10-05-17',
-                started: '2:50 AM',
-            },
-            {
-                track: 'Oh techre!',
-                length: "2:09",
-                url: 'oh-techre',
-                fileName: 'techre',
-                daw: 'LogicX',
-                date: '06-21-17',
-                started: '12:42 AM',
-            },
-            {
-                track: '1800harp',
-                length: "2:55",
-                url: '1800harp',
-                fileName: '1800harp',
-                daw: 'LogicX',
-                date: '04-16-17',
-                started: '3:17 PM',
-            },
-            {
-                track: 'Go to feather',
-                length: "1:28",
-                url: 'go2feather',
-                fileName: 'skum****',
-                daw: 'LogicX',
-                date: '07-27-17',
-                started: '1:42 AM',
-            },
-        ]
-    },
-    {
-        title: 'S&OFV4',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/s&ofv4.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv4.jpg',
-        trackList: [
-            {
-                track: 'Hello grommet',
-                length: "3:10",
-                url: 'hello-grommet',
-                fileName: 'project99',
-                daw: 'LogicX',
-                date: '06-27-17',
-                started: '4:38 PM',
-            },
-            {
-                track: 'Froggy',
-                length: "1:10",
-                url: 'froggy',
-                fileName: 'cycle psycho',
-                daw: 'LogicX',
-                date: '06-02-18',
-                started: '4:15 PM',
-            },
-            {
-                track: 'Live from the polls',
-                length: "1:08",
-                url: 'polls',
-                fileName: 'flowyhits',
-                daw: 'LogicX',
-                date: '06-19-17',
-                started: '4:44 PM',
-            },
-            {
-                track: 'BBBB',
-                length: "1:43",
-                url: 'bbbb',
-                fileName: 'wipsmalbas',
-                daw: 'LogicX',
-                date: '06-25-17',
-                started: '5:56 PM',
-            },
-            {
-                track: 'Case of the cookies',
-                length: "1:45",
-                url: 'cookies',
-                fileName: 'apple tree',
-                daw: 'LogicX',
-                date: '01-10-18',
-                started: '4:57 PM',
-            },
-            {
-                track: 'Silly Singing silicon',
-                length: "2:35",
-                url: 'silly-singing',
-                fileName: 'flunk',
-                daw: 'LogicX',
-                date: '08-28-18',
-                started: '6:22 PM',
-            },
-            {
-                track: 'Is it forever?',
-                length: "2:24",
-                url: 'is-it-forever',
-                fileName: 'flips',
-                daw: 'LogicX',
-                date: '09-13-17',
-                started: '5:02 PM',
-            },
-        ]
-    },
-    {
-        title: 'S&OFV5',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/s&ofv5.png',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv5.png',
-        trackList: [
-            {
-                track: 'Are Those Thorns?',
-                length: "1:02",
-                url: 'are-those-thorns',
-                fileName: 'dr insta',
-                daw: 'FLstudio',
-                date: '05-04-19',
-                started: '4:36 PM',
-            },
-            {
-                track: 'Barrixtown',
-                length: "1:10",
-                url: 'barrixtown',
-                fileName: 'BONKUS',
-                daw: 'FLstudio',
-                date: '08-05-19',
-                started: '12:43 PM',
-            },
-            {
-                track: 'Childs Loop',
-                length: "1:48",
-                url: 'childs-loop',
-                fileName: 'dingus',
-                daw: 'FLstudio',
-                date: '10-11-18',
-                started: '5:29 PM',
-            },
-            {
-                track: 'Cleverly Loop',
-                length: "1:43",
-                url: 'clever',
-                fileName: 'clever',
-                daw: 'FLstudio',
-                date: '11-10-18',
-                started: '5:29 PM',
-            },
-            {
-                track: 'Chix Loop',
-                length: "2:00",
-                url: 'chix',
-                fileName: 'chix',
-                daw: 'FLstudio',
-                date: '05-17-19',
-                started: '11:12 PM',
-            },
-            {
-                track: 'FL Disaster Zone',
-                length: "1:53",
-                url: 'fl-disaster',
-                fileName: 'cut90butwe',
-                daw: 'FLstudio',
-                date: '10-02-18',
-                started: '1:25 PM',
-            },
-            {
-                track: 'Crystal Envelope / Return2sender',
-                length: "4:14",
-                url: 'crystal-env',
-                fileName: 'dueling pianos',
-                daw: 'FLstudio',
-                date: '05-05-19',
-                started: '5:50 PM',
-            },
-        ]
-    },
-    {
-        title: 'S&OFV6',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/oops.png',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/oops.png',
-        trackList: [
-            {
-                track: 'Parks',
-                length: "0:20",
-                url: 'parks',
-                fileName: 'parks',
-                daw: 'FLstudio',
-                date: '04-18-19',
-                started: '1:43 PM',
-            },
-            {
-                track: 'Playing With My Hair',
-                length: "1:27",
-                url: 'playing-w-my-hair',
-                fileName: 'clair',
-                daw: 'FLstudio',
-                date: '04-13-19',
-                started: '9:32 PM',
-            },
-            {
-                track: 'Rust In Peace',
-                length: "1:43",
-                url: 'rust-in-peace',
-                fileName: 'disgus rstie',
-                daw: 'FLstudio',
-                date: '09-28-18',
-                started: '8:12 PM',
-            },
-            {
-                track: 'Kicking Space',
-                length: "1:18",
-                url: 'kicking-space',
-                fileName: 'creating space',
-                daw: 'FLstudio',
-                date: '04-13-19',
-                started: '10:57 AM',
-            },
-            {
-                track: 'Let It Fly',
-                length: "1:47",
-                url: 'let-it-fly',
-                fileName: 'destroying space',
-                daw: 'FLstudio',
-                date: '04-20-19',
-                started: '10:04 AM',
-            },
-            {
-                track: 'Liptstick Halo',
-                length: "1:31",
-                url: 'liptstick-halo',
-                fileName: 'bop',
-                daw: 'FLstudio',
-                date: '07-28-18',
-                started: '10:32 AM',
-            },
-            {
-                track: 'It Is Forever',
-                length: '?:??',
-                url: 'it-is-forever',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-
-            },
-        ]
-    },
-    {
-        title: 'S&OFV7',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/s&ofv7.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv7.jpg',
-        trackList: [
-            {
-                track: '404',
-                length: '1:00',
-                url: '404',
-                fileName: 'hat slide shift',
-                daw: 'FLstudio',
-                date: '08-24-18',
-                started: '9:20 PM',
-            },
-            {
-                track: 'Purple!',
-                length: '0:45',
-                url: 'purple',
-                fileName: 'hot spooky mess',
-                daw: 'FLstudio',
-                date: '05-31-19',
-                started: '11:54 PM',
-            },
-            {
-                track: 'Bubble Boy',
-                length: '1:54',
-                url: 'bubble-boy',
-                fileName: 'saw distort',
-                daw: 'FLstudio',
-                date: '07-05-18',
-                started: '1:48 PM',
-            },
-            {
-                track: 'Flawz',
-                length: '2:40',
-                url: 'flawz',
-                fileName: 'flaws',
-                daw: 'FLstudio',
-                date: '04-02-19',
-                started: '3:30PM',
-            },
-            {
-                track: 'Gabby',
-                length: '0:40',
-                url: 'gabby',
-                fileName: 'gabby',
-                daw: 'FLstudio',
-                date: '03-13-19',
-                started: '5:06 PM',
-            },
-            {
-                track: 'Midi Compactor',
-                length: '1:19',
-                url: 'midi-compactor',
-                fileName: 'empty',
-                daw: 'FLstudio',
-                date: '07-17-19',
-                started: '5:27 PM',
-            },
-            {
-                track: 'Rip Loop',
-                length: '3:43',
-                url: 'rip-arp',
-                fileName: 'protofunk',
-                daw: 'FLstudio',
-                date: '07-02-19',
-                started: '1:00 PM',
-            },
-        ]
-    },
-    {
-        title: 'S&OFV8',
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/s&ofv8.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/s&ofv8.jpg',
-        trackList: [
-            {
-                track: 'Rogue on 1-1',
-                length: '1:46',
-                url: 'rogue-on-1-1',
-                fileName: 'jumbles',
-                daw: 'FLstudio',
-                date: '05-03-19',
-                started: '5:12 PM',
-            },
-            {
-                track: 'Sea of Drones',
-                length: '3:34',
-                url: 'sea-of-drones',
-                fileName: 'sea of drones',
-                daw: 'FLstudio',
-                date: '06-25-19',
-                started: '10:42 PM',
-            },
-            {
-                track: 'Sine Language',
-                length: '1:42',
-                url: 'sine-lang',
-                fileName: 'drane',
-                daw: 'FLstudio',
-                date: '10-26-18',
-                started: '10:22 PM',
-            },
-            {
-                track: 'SSX Loop',
-                length: '1:08',
-                url: 'ssx-loop',
-                fileName: 'return',
-                daw: 'FLstudio',
-                date: '09-22-18',
-                started: '5:24 PM',
-            },
-            {
-                track: 'The Drs In.',
-                length: '1:18',
-                url: 'the-doctors-in',
-                fileName: 'immortal sounds',
-                daw: 'FLstudio',
-                date: '05-10-19',
-                started: '3:28 PM',
-            },
-            {
-                track: 'girard (noisey)',
-                length: '4:47',
-                url: 'girard-noisy',
-                fileName: '--',
-                daw: '--',
-                date: '--',
-                started: '--',
-            },
-        ]
-    },
-    
-    {
-        title: "IB",
-        artist: 'con',
-        longArtist: 'con',
-        releaseDate: 'Avril 4th, 2014',
-        artLink: './assets/ib.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/ib.jpg',
-        trackList: [
-            {
-                track: 'Ferox Recieved',
-                length: "3:17",
-                url: 'ferox',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Tsalal',
-                length: "1:50",
-                url: 'tsalal',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Anxious...',
-                length: "4:35",
-                url: 'anxious-oblast',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Knavish Drone (cut)',
-                length: "16:07",
-                url: 'knavish-drone',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-        ]
-    },
-    {
-        title: "Redactor",
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/redact.png',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/redact.png',
-        trackList: [
-            {
-                track: 'In a room',
-                length: "12:45",
-                url: 'on-line',
-                fileName: '--',
-                daw: 'GB/TAS',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Untitled (miccheck)',
-                length: '14:19',
-                url: 'on-line1',
-                fileName: '--',
-                daw: 'GB/TAS',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Untitled (miccheck1)',
-                length: '6:13',
-                url: 'on-line2',
-                fileName: '--',
-                daw: 'GB/TAS',
-                date: '--',
-                started: '--',
-            },
-        ]
-    },
-    {
-        title: "ECT",
-        artist: 'L. Mauro',
-        longArtist: 'Lukasz Mauro',
-        releaseDate: 'N/A',
-        artLink: './assets/sketches.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/ect.png',
-        trackList: [
-            {
-                track: 'This is forever (demo)',
-                length: "2:36",
-                url: 'this-is-forever-v1',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'HM/FL (demo)',
-                length: "7:57",
-                url: 'hmfl-demo',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'HM/FL (survivor)',
-                length: '7:57',
-                url: 'hmfl(survivor)',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-        ]
-    },
-    {
-        title: "studs",
-        artist: 'S. Finder',
-        longArtist: 'Stud Finder',
-        releaseDate: 'N/A',
-        artLink: './assets/studs.jpg',
-        webLink: 'https://forestw70.github.io/sketches-bcc-client/assets/images/studs.jpg',
-        trackList: [
-            {
-                track: 'Boiler Room',
-                length: "1:43",
-                url: 'boiler-room',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-            {
-                track: 'Chiminy Harker',
-                length: "0:43",
-                url: 'chiminy',
-                fileName: '--',
-                daw: 'FLstudio',
-                date: '--',
-                started: '--',
-            },
-        ]
-    },
-    // {
-    //     title: "BRSSF Pods",
-    //     artist: "Burner B",
-    //     longArtist: "DJ Burna B",
-    //     releaseDate: "September 30th, 2018",
-    //     artLink: "./assets/pods.jpg",
-    //     webLink: "https://burnthis.space/assets/images/pods.jpg",
-    //     trackList: [
-    //         {
-    //             track: 'Soundscape Source Feed vol. 1',
-    //             length: "33:33",
-    //             url: 'pod1-198',
-    //             fileName: 'brssfv1',
-    //             daw: 'FLstudio',
-    //             date: '09-30-18',
-    //             started: '--',
-    //         },
-    //         {
-    //             track: 'Soundscape Source Feed vol. 2',
-    //             length: "55:08",
-    //             url: 'pod2-112',
-    //             fileName: 'brssfv2',
-    //             daw: 'FLstudio',
-    //             date: '12-26-18',
-    //             started: '--',
-    //         },
-    //         {
-    //             track: 'Soundscape Source Feed vol. 3',
-    //             length: "40:27",
-    //             url: 'pod3-160',
-    //             fileName: 'brssfv3',
-    //             daw: 'FLstudio',
-    //             date: '03-01-19',
-    //             started: '--',
-    //         }
-    //     ]
-    // }
-]
-
-// directory
-const headSwap = document.getElementById("wellHeyThere");
-const audioPlayer = document.getElementById("player");
-const currAlbumPic = document.getElementById("albumPic");
-const currTrackName = document.getElementById("trackName");
-const currArtistName = document.getElementById("artistName");
-const currTrackTime = document.getElementById("trackTime");
-const currAlbumName = document.getElementById("albumName");
-const queueListDump = document.getElementById("queueList");
-const qToolCont = document.getElementById("qTools")
-const discoContainer = document.getElementById("disco");
-const listView = document.getElementById("list");
-const listDump = document.getElementById("listDump");
-const optionsContainer = document.getElementById("options");
-const testBtn = document.getElementById("testBtn");
-
-
-// SONGQUEUE CLASS
-// 
-function SongQueue(pushbox = []) {
-    this.allTheSongs = pushbox;
-    this.activeQueue = [];
-
-    // all song list for sorting
-    this.populateSongList = () => {
-        const listSort = getDefaultList();
-        return this.allTheSongs.push(listSort);
-    }
-    this.getSongList = () => {
-        return this.allTheSongs[0];
-    }
-
-    this.changeSongList = (newSorted) => {
-        this.allTheSongs = [];
-
-        return this.allTheSongs.push(newSorted);
-    }
-
-    // queue actions
-    this.add2queue = (el) => {
-        const item = {
-            songTitle: el.title,
-            songLength: el.length,
-            songEp: el.ep,
-            songArtist: el.artist,
-            songUrl: el.url,
-            songArt: el.art,
-        }
-        return this.activeQueue.push(item)
-    }
-
-    this.add2FrontOfQueue = (el) => {
-        const item = {
-            songTitle: el.title,
-            songLength: el.length,
-            songEp: el.ep,
-            songArtist: el.artist,
-            songUrl: el.url,
-            songArt: el.artLink,
-        }
-        this.activeQueue.splice(1, 0, item);
-        
-        const qButton = document.createElement("button");
-        qButton.addEventListener("click", handleQClick);
-        qButton.type = "button"
-        qButton.innerText = `${el.url} ~ ${el.ep}`;
-        qButton.id = el.url + " " + el.ep;
-
-        const firstChild = queueListDump.firstChild
-        queueListDump.insertBefore(qButton, firstChild);
-    }
-
-    this.createQItem = (el) => {
-        const qId = el.url + " " + el.ep
-        const qButton = document.createElement("button");
-        qButton.addEventListener("click", handleQClick);
-
-        qButton.type = "button"
-        qButton.innerText = `${el.url} ~ ${el.ep}`;
-        qButton.id = qId;
-        if (this.doesButtonExist(qId) === 1) {
-            console.log("already in queue!")
-            return;
-        } else {
-            queueListDump.appendChild(qButton);
-        }
-    }
-
-    this.removeQueueButton = (btnId) => {
-        document.getElementById(btnId).remove();
-        return;
-    }
-
-    this.removePlaceInQueue = (trackId) => {
-        const url = trackId.split(" ")[0]
-        const trackIdx = this.activeQueue.findIndex(item => item.songUrl === url)
-        const removed = this.activeQueue.splice(trackIdx, 1)
-
-        this.removeQueueButton(trackId);
-        console.log(removed);
-    }
-
-    this.removeFirstFromQueue = () => {
-        const first = this.activeQueue[1];
-        const searchBtn = first.songUrl + " " + first.songEp;
-        this.removeQueueButton(searchBtn)
-        return this.activeQueue.shift();
-    }
-
-    this.findButtonSearch = (songUrl) => {
-        const getId = this.activeQueue.find(track => track.songUrl === songUrl)
-        const searchId = getId.songUrl + " " + getId.songEp
-        return searchId;
-    }
-
-    this.doesButtonExist = (songObj) => {
-        const searchId = songObj.url + " " + songObj.ep;
-        const buttonSearch = document.getElementById(searchId)
-        if (buttonSearch === null) {
-            return 0
-        }
-        return 1;
-    }
-
-    this.grabNext = () => {
-        return this.activeQueue[0];
-    }
-
-    this.grabLength = () => {
-        return this.activeQueue.length;
-    }
-
-    this.findSong = (url) => {
-        const songName = url;
-        // const ep = id.split('~')[1];
-
-        let item = this.allTheSongs[0].find(song => song.url === songName)
-        return item;
-    }
-
-    this.filterPlayer = () => {
-        let wiper = this.activeQueue[0]
-        if (this.activeQueue.length >= 2) {
-            wiper = this.activeQueue[1]
-        }
-        const songHome = `https://forestw70.github.io/sketches-bcc-client/assets/music/${wiper.songUrl}.mp3`
-        const newTitle = `${wiper.songArtist} ~ ${wiper.songUrl}`
-
-        // clear page title and replace with song info
-        clearPlayer();
-        document.title = newTitle;
-        headSwap.innerText = wiper.songArtist + " " + wiper.songUrl + ".mp3 '" + wiper.songTitle + "' " + wiper.songEp + " " + wiper.songLength;
-        currArtistName.innerText = wiper.songArtist;
-        currTrackName.innerText = `"${wiper.songUrl}"`;
-        currAlbumName.innerText = wiper.songEp;
-        currAlbumPic.src = wiper.songArt;
-        audioPlayer.src = songHome;
-        return console.log("Next track loaded!");
-    }
-}
-const songQueue = new SongQueue();
-
-// class to control individual song actions 
-// (artist, title, ep, art, url, length, released, artistlong, ogFileName, dawUsed, dateCreated, timeCreated)
-function Song(artist, title, ep, art, url, length, released, long, ogFile, daw, startDate, startTime) {
-    this.artist = artist;
-    this.artistLong = long;
-    this.title = title;
-    this.ep = ep;
-    this.art = art;
-    this.url = url;
-    this.length = length;
-    this.released = released;
-    this.ogFile = ogFile;
-    this.daw = daw;
-    this.startDate = startDate;
-    this.startTime = startTime;
-
-    this.createListSongRow = (trackNum) => {
-        const trackNumber = document.createElement("span");
-        const trackName = document.createElement('span');
-        const trackLength = document.createElement('span');
-        const albumTitle = document.createElement('span');
-        const trackArtist = document.createElement('span');
-        const trackUrl = document.createElement('span');
-        const trackOgFile = document.createElement('span');
-        const trackDate = document.createElement('span');
-        const trackTimeStamp = document.createElement('span');
-        // set info
-        trackNumber.innerText = trackNum;
-        trackName.innerText = this.title;
-        trackLength.innerText = this.length
-        albumTitle.innerText = this.ep;
-        trackArtist.innerText = this.artist;
-        trackUrl.innerText = this.url;
-        trackOgFile.innerText = this.ogFile;
-        trackDate.innerText = this.startDate;
-        trackTimeStamp.innerText = this.startTime;
-        // row button
-        const trackRow = document.createElement("button");
-        trackRow.type = "button";
-        trackRow.classList.add('song-list-row');
-        trackRow.dataset.songId = this.title + '~' + this.ep;
-        trackRow.dataset.albumurl = this.art;
-        trackRow.dataset.title = this.title;
-        trackRow.dataset.length = this.length;
-        trackRow.dataset.epname = this.ep;
-        trackRow.dataset.artist = this.artist;
-        trackRow.dataset.url = this.url;
-        trackRow.dataset.ogFile = this.ogFile;
-        trackRow.dataset.dCreated = this.startDate;
-        trackRow.dataset.tCreated = this.startTime;
-        trackRow.addEventListener("click", handleTrackSelect);
-        // append
-        trackRow.appendChild(trackNumber);
-        trackRow.appendChild(trackName);
-        trackRow.appendChild(trackLength);
-        trackRow.appendChild(albumTitle);
-        trackRow.appendChild(trackArtist);
-        trackRow.appendChild(trackUrl);
-        trackRow.appendChild(trackOgFile);
-        trackRow.appendChild(trackDate);
-        trackRow.appendChild(trackTimeStamp);
-        listDump.appendChild(trackRow);
-    }
-
-    this.createAlbumSongRow = (trackNum) => {
-        const trackRow = document.createElement("button");
-        trackRow.type = "button";
-        trackRow.classList.add("t-row");
-        trackRow.dataset.url = this.url;
-        trackRow.dataset.length = this.length;
-        trackRow.dataset.epname = this.ep;
-        trackRow.dataset.albumurl = this.art;
-        trackRow.dataset.artist = this.artist;
-        trackRow.dataset.title = this.title;
-        trackRow.addEventListener("click", handleTrackSelect);
-        //
-        const tTitle = document.createElement("span");
-        const tTime = document.createElement("span");
-        const tNum = document.createElement("span");
-        tTitle.classList.add("album-track");
-        tTime.classList.add("track-time");
-        tNum.classList.add("track-num");
-        tTitle.innerText = this.title;
-        tTime.innerText = this.length;
-        tNum.innerText = trackNum;
-        //
-        trackRow.appendChild(tNum);
-        trackRow.appendChild(tTitle);
-        trackRow.appendChild(tTime);
-        return trackRow;
-    }
-}
-
-// UTILITY AND OTHER FEATURES
-//
-
-// clear audio player
-const clearPlayer = () => {
-    currArtistName.innerText = "";
-    currTrackName.innerText = "";
-    currAlbumName.innerText = "No track queued!";
-    currAlbumPic.src = "https://picsum.photos/25";
-    audioPlayer.src = "https://forestw70.github.io/sketches-bcc-client/assets/music/mario3.mp3";
-    document.title = "Sketches";
-    headSwap.innerText = "sum' sketches"
-}
-
-// top of page button
-const topOfPageBtn = document.getElementById("returnToTop")
-topOfPageBtn.addEventListener("click", () => {
-    document.body.scrollTop = 0;
-    document.documentElement.scrollTop = 0;
-})
-
-window.onscroll = () => {
-    if (document.body.scrollTop > 200 || document.documentElement.scrollTop > 200) {
-        topOfPageBtn.style.display = "flex";
-    } else {
-        topOfPageBtn.style.display = "none";
-    }
-};
-
-// autoplay option button
-let autoPlayOn = false;
-const autoPlayButton = document.getElementById("autoPlayOption")
-autoPlayButton.addEventListener("click", (e) => {
-    e.preventDefault();
-    if (!autoPlayOn) {
-        autoPlayButton.innerText = "~expirimental autoplay~ -- ON."
-        autoPlayButton.classList.add("opt-on")
-        autoPlayOn = true;
-        return;
-    }
-    autoPlayOn = false;
-    autoPlayButton.innerText = "~expirimental autoplay~ -- OFF."
-    autoPlayButton.classList.remove("opt-on")
-})
-
-audioPlayer.addEventListener("ended", () => {
-    if (!autoPlayOn) {
-        console.log("auto play is not on.")
-        return;
-    }
-    const qtime = songQueue.grabLength()
-    console.log(qtime)
-    if (qtime > 1) {
-        songQueue.filterPlayer();
-        songQueue.removeFirstFromQueue();
-        audioPlayer.play();
-    } else if (qtime <= 1) {
-        console.log('no more items in queue.')
-    }
-})
-
-
-//  TRACK SELECT
-// 
-const handleTrackSelect = (e) => {
-    e.preventDefault();
-    const lookObj = {
-        title: e.currentTarget.dataset.title,
-        length: e.currentTarget.dataset.length,
-        ep: e.currentTarget.dataset.epname,
-        artist: e.currentTarget.dataset.artist,
-        url: e.currentTarget.dataset.url,
-        art: e.currentTarget.dataset.albumurl,
-    }
-    const currQueueTime = songQueue.grabLength();
-    if (currQueueTime === 0) {
-        songQueue.add2queue(lookObj);
-        songQueue.filterPlayer();
-    } else {
-        if (songQueue.doesButtonExist(lookObj) === 0) {
-            songQueue.add2queue(lookObj);
-            songQueue.createQItem(lookObj);
-        } else {
-            window.alert("song already in queue!")
-        }
-    }
-}
-
-
-//  ALBUM VIEW
-//  
-const showAlbumView = () => {
-    localStorage.setItem("currentView", "album")
-
-    discoContainer.innerText = ""
-    listDump.innerText = '';
-    listView.classList.add("hide");
-    optionsContainer.classList.add("hide");
-    discoContainer.classList.remove("hide");
-
-
-    sketches.map(album => {
-        const idvAlbum = buildAlbumCont(album);
-        discoContainer.appendChild(idvAlbum);
-    })
-
-}
-
-const buildAlbumCont = (albumInfo) => {
-    const albumContainer = document.createElement("article");
-    const infoContainer = document.createElement("div");
-    const trackContainer = document.createElement("div");
-    const albumTitle = document.createElement("h3");
-    const albumArtist = document.createElement("h2");
-    const albumRelease = document.createElement("p");
-    const albumArt = document.createElement("img");
-    //
-    albumContainer.classList.add("disco-item");
-    infoContainer.classList.add("info-container");
-    trackContainer.classList.add("track-container");
-    trackContainer.id = albumInfo.title;
-    albumTitle.innerText = `"${albumInfo.title}"`;
-    albumArtist.innerText = albumInfo.artist;
-    albumRelease.innerText = albumInfo.releaseDate;
-    albumArt.src = albumInfo.webLink;
-    //
-    infoContainer.appendChild(albumArt)
-    infoContainer.appendChild(albumArtist)
-    infoContainer.appendChild(albumTitle)
-    infoContainer.appendChild(albumRelease)
-    albumContainer.appendChild(infoContainer)
-    //
-    albumInfo.trackList.map((idvTrack, idx) => {
-        const albumSong = new Song(
-            albumInfo.artist,
-            idvTrack.track,
-            albumInfo.title,
-            albumInfo.webLink,
-            idvTrack.url,
-            idvTrack.length,
-            albumInfo.releaseDate,
-            albumInfo.artistLong,
-            idvTrack.fileName,
-            idvTrack.daw,
-            idvTrack.date,
-            idvTrack.started
-        );
-        const nextRow = albumSong.createAlbumSongRow(idx + 1);
-        trackContainer.appendChild(nextRow);
-    })
-    albumContainer.appendChild(trackContainer)
-    return albumContainer;
-}
-
-
-
-// SONG VIEW
-// 
-const showSongView = (sortedSongList) => {
-    localStorage.setItem("currentView", "list")
-    listDump.innerText = "";
-    discoContainer.classList.add('hide');
-    optionsContainer.classList.add("hide");
-    listView.classList.remove("hide");
-    // createHeaderRow();
-    sortedSongList.map((song, idx) => {
-        let track = new Song(
-            song.artist,
-            song.title,
-            song.ep,
-            song.artLink,
-            song.url,
-            song.length,
-            song.releaseDate,
-            song.artistLong,
-            song.ogFileName,
-            song.dawUsed,
-            song.dateCreated,
-            song.timeCreated
-        )
-        track.createListSongRow(idx)
-    })
-}
-
-const getDefaultList = () => {
-    let songId = 0;
-    const sorted = [];
-    sketches.map(album => {
-        album.trackList.map(song => {
-            const listItem = {
-                trackNumber: songId,
-                title: song.track,
-                ep: album.title,
-                length: song.length,
-                url: song.url,
-                artLink: album.webLink,
-                artist: album.artist,
-                artistLong: album.artistLong,
-                releaseDate: album.releaseDate,
-                ogFileName: song.fileName,
-                dawUsed: song.daw,
-                dateCreated: song.date,
-                timeCreated: song.started,
-            }
-            sorted.push(listItem)
-            songId++
-        })
-    })
-    return sorted;
-}
-
-
-// OPTIONS VIEW
-// 
-const showOptionView = () => {
-    listView.classList.add('hide');
-    listDump.innerText = '';
-    discoContainer.classList.add('hide');
-    optionsContainer.classList.remove('hide');
-    localStorage.setItem("currentView", "option")
-
-}
-
-
-
-// SORT FUNCTIONALITY
-// 
-const splitLength = (time) => {
-    // "2:32"
-    if (time === "?:??") return 1;
-
-    const els = time.split(":");
-    return new Number((els[0] * 60) + els[1])
-}
-
-const splitDate = (date) => {
-    // "6-23-18"
-    if (date === "--") return 1;
-
-    const els = date.split("-");
-    const utcDate = new Date(Date.UTC(els[2], els[0] - 1, els[1]))
-    return utcDate.getTime();
-}
-
-const splitTimeSpamp = (ts) => {
-    // "12:05 PM"
-    if (ts === "--") return 1;
-
-    const els = ts.split(" ");
-    let mins = els[0].split(":")[0];
-    const secs = els[0].split(":")[1];
-    if (els[1] === "PM") {
-        mins = mins + 12;
-    }
-
-    const totalSecs = new Number((mins * 60) + secs);
-    return totalSecs;
-}
-
-
-// const doMath = (item, split) => {
-//     if (item === "--" || item === "?:??") return 1;
-//     if (split === "date") {
-//         const els = item.split("-");
-//         const utcDate = new Date(Date.UTC(els[2], els[0] - 1, els[1]))
-//         return utcDate.getTime();
-//     } else {
-//         const minSec = item.split(":");
-//         let mins = minSec[0];
-//         let secs = minSec[1];
-//         if (split === "time") {
-//             let ampm = secs.split(" ")[1];
-//             secs = secs.split(" ")[0];
-//             if (ampm === "PM") mins = mins + 12;
-//         }
-//         const totalTime = new Number((mins * 60) + secs);
-//         return totalTime;
-//     }
-// }
-
-// const constructSort = (type, param, split) => {
-
-//     const currentList = songQueue.getSongList();
-//     let nextSort = [];
-//     const pKey = param;
-
-//     if (type === "num") {
-//         if (split == 0) {
-//             nextSort = currentList.sort((a, b) => {
-//                 let el1 = a.pKey;
-//                 let el2 = b.pKey;
-//                 return el2 - el1;
-//             })
-//             return nextSort;
-//         } else {
-//             nextSort = currentList.sort((a, b) => {
-//                 let el1 = doMath(a.pKey, split);
-//                 let el2 = doMath(b.pKey, split);
-//                 return el1 - el2;
-//             })
-//             return nextSort;
-//         }
-//     } else if (type === "string") {
-//         nextSort = currentList.sort((a, b) => {
-//             if (a.pKey === "--" || b.pKey === "--") return 1;
-//             let el1 = a.title.toUpperCase();
-//             let el2 = b.title.toUpperCase();
-
-//             if (el1 > el2) {
-//                 return 1
-//             } else if (el2 > el1) {
-//                 return -1
-//             } else {
-//                 return 0
-//             }
-//         })
-//         return nextSort;
-//     } else {
-//         console.log("doh");
-//     }
-// }
-
-
-// CHANGE SORT 
-// 
-const changeSortedSongList = (sortBy) => {
-    const currentSort = songQueue.getSongList();
-    let newSort;
-    switch (sortBy) {
-        case "trkNum":
-            console.log("sorting by track number (useless)..")
-            newSort = currentSort.sort((a, b) => {
-                return b.trackNumber - a.trackNumber;
-            })
-            // newSort = constructSort('num', 'trackNumber', 0)
-            break;
-
-        case "trkNme":
-            console.log("sorting by track title..");
-            newSort = currentSort.sort((a, b) => {
-                let el1 = a.title.toUpperCase();
-                let el2 = b.title.toUpperCase();
-                if (el1 > el2) {
-                    return 1
-                }
-                if (el1 < el2) {
-                    return -1
-                }
-                return 0;
-            })
-
-
-            break;
-
-        case "trkLen":
-            console.log("sorting by track length..")
-            newSort = currentSort.sort((a, b) => {
-                let el1 = splitLength(a.length);
-                let el2 = splitLength(b.length);
-                return el1 - el2;
-
-            })
-            break;
-
-        case "epTtl":
-            console.log("sorting by ep name..")
-            newSort = currentSort.sort((a, b) => {
-                let el1 = a.ep.toUpperCase();
-                let el2 = b.ep.toUpperCase();
-                if (el1 > el2) {
-                    return 1
-                }
-                if (el1 < el2) {
-                    return -1
-                }
-                return 0;
-            })
-
-            break;
-
-        case "artist":
-            console.log("sorting by track artist..")
-            newSort = currentSort.sort((a, b) => {
-                let el1 = a.artist.toUpperCase();
-                let el2 = b.artist.toUpperCase();
-                if (el1 > el2) {
-                    return 1
-                }
-                if (el1 < el2) {
-                    return -1
-                }
-                return 0;
-            })
-
-            break;
-
-        case "trkUrl":
-            console.log("sorting by track url..")
-            newSort = currentSort.sort((a, b) => {
-                let el1 = a.url.toUpperCase();
-                let el2 = b.url.toUpperCase();
-                if (el1 > el2) {
-                    return 1
-                }
-                if (el1 < el2) {
-                    return -1
-                }
-                return 0;
-
-            })
-            break;
-
-        case "trkOg":
-            console.log("sorting by file name..")
-            newSort = currentSort.sort((a, b) => {
-                if (a.fileName === "--" || b.fileName === "--") return 0;
-
-                let el1 = a.ogFileName.toUpperCase();
-                let el2 = b.ogFileName.toUpperCase();
-                if (el1 > el2) {
-                    return -1
-                }
-                if (el1 < el2) {
-                    return 1
-                }
-                return 0;
-            })
-            break;
-
-        case "trkDte":
-            console.log("sorting by date created..")
-            newSort = currentSort.sort((a, b) => {
-                let el1 = splitDate(a.dateCreated);
-                let el2 = splitDate(b.dateCreated);
-                return el1 - el2;
-
-            })
-            break;
-
-        case "trkTs":
-            console.log("sorting by timestamp..")
-            newSort = currentSort.sort((a, b) => {
-                let el1 = splitTimeSpamp(a.timeCreated);
-                let el2 = splitTimeSpamp(b.timeCreated);
-                return el1 - el2;
-
-            })
-            break;
-
-        default:
-            console.log(";)")
-    }
-    songQueue.changeSongList(newSort);
-}
-
-// localstorage
-window.localStorage.setItem("view", "albumView")
-
-
-// queue remove
-const modal = document.getElementById("qPrompt");
-const modalDump = document.getElementById("modalDump");
-const modalSong = document.getElementById("modalSong");
-const closeModalBtn = document.getElementById("closeModal");
-const removeBtn = document.getElementById("removeThisSong");
-const upNextBtn = document.getElementById("upNextBtn");
-const nowPlayingInfo = document.getElementById("nowPlayingInfo");
-
-
-const handleCurrentTrackClick = (url) => {
-    
-    const data4Modal = songQueue.findSong(url)
-    modalDump.innerText = `${url}=${JSON.stringify(data4Modal, null, 2)}`;
-    removeBtn.style.display = "none"
-    upNextBtn.style.display = "none"
-    modal.style.display = "block";
-}
-
-const handleQClick = (e) => {
-    e.preventDefault();
-    const searchUrl = e.target.id.split(" ")[0]
-    const data4Modal = songQueue.findSong(searchUrl)
-    modalDump.innerText = `${searchUrl}=${JSON.stringify(data4Modal, null, 2)}`;
-    removeBtn.dataset.whichSong = searchUrl
-    upNextBtn.dataset.whichSong = searchUrl
-    modal.style.display = "block";
-    removeBtn.style.display = "block"
-    upNextBtn.style.display = "block"
-}
-
-const closeModal = () => {
-    modal.style.display = "none";
-    removeBtn.dataset.whichSong = ""
-    modalDump.innerText = ""
-    removeBtn.style.display = "none"
-    upNextBtn.style.display = "none"
-}
-
-nowPlayingInfo.addEventListener("click", (e) => {
-    e.preventDefault();
-    const displayFor = currTrackName.innerText.split('"')[1]
-    handleCurrentTrackClick(displayFor)
-})
-
-upNextBtn.addEventListener("click", (e) => {
-    e.preventDefault();
-    const search = e.target.dataset.whichSong
-    const btnSearch = songQueue.findButtonSearch(search);
-    songQueue.removePlaceInQueue(btnSearch);
-
-    const songItem = songQueue.findSong(search);
-    songQueue.add2FrontOfQueue(songItem);
-    
-    closeModal();
-})
-
-removeBtn.addEventListener("click", (e) => {
-    e.preventDefault();
-    const search = e.target.dataset.whichSong
-    const btnSearch = songQueue.findButtonSearch(search)
-    songQueue.removePlaceInQueue(btnSearch)
-    closeModal();
-})
-
-closeModalBtn.addEventListener("click", (e) => {
-    e.preventDefault();
-    closeModal();
-})
-
-window.onclick = (event) => {
-    if (event.target == modal) {
-        modalDump.innerText = ""
-        modal.style.display = "none";
-    }
-}
-
-// BUTTONS
-// 
-const toggleQueue = document.getElementById("toggleShowQueue")
-const albumViewBtn = document.getElementById("viewAlbums")
-const songViewBtn = document.getElementById("viewSongs")
-const optViewBtn = document.getElementById("viewOptions")
-const nextTrackBtn = document.getElementById("nextTrack")
-const playButton = document.getElementById("pause-play")
-const queueTip = document.getElementById("qTip")
-const icon = document.getElementById("pp")
-const queueicon = document.getElementById("sq")
-
-// toggle show queue
-toggleQueue.addEventListener("click", () => {
-    if (!queueListDump.classList.contains("menu-hide")) {
-
-        queueicon.classList.remove("glyphicon-menu-right")
-        queueicon.classList.add("glyphicon-menu-down")
-        queueListDump.classList.toggle('menu-hide')
-        // queue hidden
-        queueTip.innerText = "Queue Hidden."
-
-    } else {
-
-        queueListDump.classList.toggle('menu-hide')
-        queueicon.classList.remove("glyphicon-menu-down")
-        queueicon.classList.add("glyphicon-menu-right")
-        queueTip.innerText = "Queue:"
-    }
-})
-
-// change views
-albumViewBtn.addEventListener("click", () => {
-    if (!albumViewBtn.classList.contains("curr-view")) {
-        albumViewBtn.classList.add("curr-view")
-        optViewBtn.classList.remove("curr-view")
-        songViewBtn.classList.remove("curr-view")
-        showAlbumView();
-    }
-})
-songViewBtn.addEventListener("click", () => {
-    if (!songViewBtn.classList.contains("curr-view")) {
-        songViewBtn.classList.add("curr-view")
-        albumViewBtn.classList.remove("curr-view")
-        optViewBtn.classList.remove("curr-view")
-        showSongView(songQueue.getSongList());
-    }
-})
-optViewBtn.addEventListener("click", () => {
-    if (!optViewBtn.classList.contains("curr-view")) {
-        optViewBtn.classList.add("curr-view")
-        albumViewBtn.classList.remove("curr-view")
-        songViewBtn.classList.remove("curr-view")
-        showOptionView();
-    }
-})
-
-// Audio Player Buttons
-playButton.addEventListener("click", () => {
-    const isLoadedCheck = audioPlayer.src.split(".").pop()
-    console.log(isLoadedCheck)
-    if (isLoadedCheck !== "mp3") {
-        return window.alert("Please queue up next song.")
-    }
-    if (audioPlayer.paused) {
-        
-        icon.classList.remove("glyphicon-play");
-        icon.classList.add("glyphicon-pause");
-        audioPlayer.play();
-    } else {
-        
-        audioPlayer.pause();
-        icon.classList.remove("glyphicon-pause");
-        icon.classList.add("glyphicon-play");
-    }
-})
-
-nextTrackBtn.addEventListener("click", () => {
-
-    const queueLength = songQueue.grabLength();
-    icon.classList.remove("glyphicon-pause");
-    icon.classList.add("glyphicon-play");
-    console.log(`You have ${queueLength - 2} items left in your queue!`);
-    if (queueLength <= 1) {
-        clearPlayer();
-        console.log("nothing queued!")
-    } else {
-        songQueue.filterPlayer();
-        songQueue.removeFirstFromQueue();
-        icon.classList.add("glyphicon-pause");
-        icon.classList.remove("glyphicon-play");
-        audioPlayer.play();
-    }
-})
-
-
-// songview header bar sort buttons
-const headButtons = document.getElementById("headerRow").querySelectorAll("span");
-headButtons.forEach(button => {
-    button.addEventListener("click", async (e) => {
-        e.preventDefault();
-        if (e.target.classList.contains("list-reverse")) {
-            let list = songQueue.getSongList();
-            e.target.classList.remove("list-reverse")
-            return showSongView(list.reverse());
-        }
-
-        await changeSortedSongList(e.target.dataset.sortBy)
-        showSongView(songQueue.getSongList())
-        e.target.classList.add("list-reverse");
-
-    })
-})
-
-// testBtn.addEventListener("click", (e) => {
-//     e.preventDefault();
-//     changeSortedSongList("trkNme");
-// })
-
-window.onload = () => {
-    songQueue.populateSongList();
-    const currView = localStorage.getItem("currentView")
-    if (currView === "list") {
-        songViewBtn.classList.add("curr-view")
-        showSongView(songQueue.getSongList());
-    } else {
-        console.log(currView)
-        albumViewBtn.classList.add("curr-view")
-        showAlbumView();
-    }
-}
-
