+++ index.html
@@ -173,4 +173,4 @@
 
 </html>
 
-<script src="assets/scripts/scripts.js"></script>
+<script type="module" src="assets/js/index.js"></script>
